Add machine translation workflow setting in tab UI
Add a checkbox to set `isWorkflowMT` boolean, passed with `createTask`
to allow for creating a job using the machine translation workflow
instead of the default workflow for each locale.

Will want to improve the display text/UI, and possibly accept an array
of desired workflows to choose from instead of just MT as a bool.