Use workflowOptions param to set workflow
- Adds an optional `workflowOptions` param to the config passed to
TranslationsTab from the studio, which accepts an array of workflows.
Workflows are expected to include a `workflowUid`, which can be found
in the Smartling project's UI, and a `workflowName` for display title.
- The workflow select input only shows if at least one workflow option
is passed, and otherwise the additional work of setting the workflow
and locales in task creation isn't done
- The selected workflowUid string (if present) is passed as a param
with the rest of the context to the adapter's `createTask` function
- This replaces the `isWorkflowMT` boolean, allowing user to pass in
one or more specific workflow uids and names.