Accept localeIdAdapter for custom ID transforms
Locales may have different IDs in the translation vendor and in Sanity.
This adds an option to the plugin that accepts a function
`localeIdAdapter`, which will be passed the translation vendor locale
id and is expected to return the Sanity locale id. The translation
vendor id will be used for `getTranslation`, but then the Sanity id is
used for `context.importTranslation`.

Using the `localeIdAdapter` here removes the need to implement it
across multiple adapters and allows default adapter configs to be used
when the only required change is the locale ids.

Example implementation with sanity-plugin-studio-smartling adapter:

```js
// deskStructure.js
import { defaultFieldLevelConfig } from 'sanity-plugin-studio-smartling'
import { TranslationsTab } from 'sanity-translations-tab'

const locales = {}
supportedLanguages.forEach(({ smartlingId, id }) => {
  locales[smartlingId || id] = id
})

export const getDefaultDocumentNode = (props) => {
  return S.document().views([
    S.view.form(),
    S.view.component(TranslationsTab).title('Smartling').options({
     ...defaultFieldLevelConfig,
     localeIdAdapter: (smartlingId) => locales[smartlingId]
    })
  ])
}
```