Disable button during import
While the import is in progress, we want this visually reflected on the
button so the user knows that it's working in the background and doesn't
try to click again.

`importFile` and `handleImport` need to be `async` employ `await` so
that `setIsBusy(false)` isn't executed until the task is complete.

While loading (`setIsBusy === true`), the button is disabled and the
content changes to `Loading...` with no icon, because the button doesn't
have a lot of room.

Also updates the button to be disabled when there's no `progress` at all
or if it's `0`.