Show link to external job, if present
This update accepts `linkToVendorTask` as an additional detail returned
with a `TranslationTask`, which should be a link to view the current
translation task on the vendor's platform.
If there is a translation task and it includes `linkToVendorTask`, then
a link to "View Job" will show in the translation tab so users can
easily get to the job on the translation vendor platform to perform  additional tasks or get additional info.